requirements.txt
fastapi[standard]==0.115.11
alembic==1.15.1
SQLAlchemy==2.0.39
sqlmodel==0.0.24
uvicorn==0.29.0
python-jose[cryptography]==3.4.0
pydantic-settings==2.8.1
psycopg2-binary==2.9.10
python-dotenv==1.1.0
passlib[bcrypt]==1.7.4